+++ test/json_utils_double.h
@@ -6,17 +6,17 @@
 #include <iostream>
 #include "rossy_utils/io/json.hpp"
 
-Eigen::VectorXf json_list_to_eigen(const nlohmann::json &j){
-    Eigen::VectorXf vector(j.size());
+Eigen::VectorXd json_list_to_eigen(const nlohmann::json &j){
+    Eigen::VectorXd vector(j.size());
     int element_index=0;    
     for (const auto& element : j) {
-        vector(element_index++) = (float)element;
+        vector(element_index++) = (double)element;
     }
     return vector;
 }
 
-Eigen::MatrixXf json_listoflist_to_eigenmat(const nlohmann::json &j){
-    Eigen::MatrixXf mat(j[0].size(),j.size());
+Eigen::MatrixXd json_listoflist_to_eigenmat(const nlohmann::json &j){
+    Eigen::MatrixXd mat(j[0].size(),j.size());
     size_t element_index=0;
     for (const auto& element : j) {
         mat.col(element_index++) = json_list_to_eigen(element);
@@ -25,9 +25,9 @@ Eigen::MatrixXf json_listoflist_to_eigenmat(const nlohmann::json &j){
 }
 
 void json_listoflist_to_vecofeigen(const nlohmann::json &j, 
-                    std::vector<Eigen::VectorXf> &vector,
+                    std::vector<Eigen::VectorXd> &vector,
                     bool rowwise = true){
-    Eigen::MatrixXf mat =json_listoflist_to_eigenmat(j);
+    Eigen::MatrixXd mat =json_listoflist_to_eigenmat(j);
     vector.clear();
     if(rowwise){
         for(int i(0); i<mat.rows(); ++i)
